% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_md_line}
\alias{add_md_line}
\title{Append lines to markdown state}
\usage{
add_md_line(state, x)
}
\arguments{
\item{state}{Internal state environment.}

\item{x}{Character vector.}
}
\value{
Invisibly returns NULL.
}
\description{
Append lines to markdown state
}
\keyword{internal}
