% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_md_text_block}
\alias{add_md_text_block}
\title{Append a fenced text block to markdown state}
\usage{
add_md_text_block(state, x)
}
\arguments{
\item{state}{Internal state environment.}

\item{x}{Character vector.}
}
\value{
Invisibly returns NULL.
}
\description{
Append a fenced text block to markdown state
}
\keyword{internal}
