% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birtir-package.R
\docType{package}
\name{birtir-package}
\alias{birtir}
\alias{birtir-package}
\title{birtir: Render Analysis Scripts to Plain Markdown Reports}
\description{
\code{birtir} runs plain \code{.R} scripts and turns them into structured Markdown
reports, saving plots and tables alongside the generated \code{.md} file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/auv2/birtir}
  \item Report bugs at \url{https://github.com/auv2/birtir/issues}
}

}
\author{
\strong{Maintainer}: Auðun Valborgarson \email{your@email.com}

}
\keyword{internal}
