% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-labels.R
\name{description_labels}
\alias{description_labels}
\title{Create output labels for birtir summaries}
\usage{
description_labels(
  table = "Table",
  figure = "Figure",
  variable = "Variable",
  summary = "Summary",
  type = "Type",
  formula = "Formula",
  missing = "Missing"
)
}
\arguments{
\item{table}{Label for tables.}

\item{figure}{Label for figures.}

\item{variable}{Label for variables.}

\item{summary}{Label for summaries.}

\item{type}{Label for variable type.}

\item{formula}{Label for formulas.}

\item{missing}{Label for missingness.}
}
\value{
A \code{description_labels} object.
}
\description{
Create output labels for birtir summaries
}
