% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-helpers.R
\name{inline_bar}
\alias{inline_bar}
\title{Inline bar chart}
\usage{
inline_bar(values, width = 20, digits = 1)
}
\arguments{
\item{values}{Named numeric vector of values, such as group means.}

\item{width}{Width of the inline plot in characters.}

\item{digits}{Number of decimal places for printed values.}
}
\value{
A \code{birtir_inline_bar} object.
}
\description{
Inline bar chart
}
\examples{
inline_bar(c(g0 = 49.1, g1 = 51.4))
x <- c(41, 53, 49, 52, NA, 47)
group <- factor(c("a", "a", "b", "b", "a", "b"))
inline_bar(tapply(x, group, mean, na.rm = TRUE))
}
