% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-helpers.R
\name{inline_boxplot}
\alias{inline_boxplot}
\title{Inline boxplot}
\usage{
inline_boxplot(x, width = 30)
}
\arguments{
\item{x}{Numeric vector.}

\item{width}{Width of the inline plot in characters.}
}
\value{
A \code{birtir_inline_boxplot} object.
}
\description{
Inline boxplot
}
\examples{
inline_boxplot(c(1, 2, 3, 4, 5))
x <- c(41, 53, 49, 52, NA, 47)
inline_boxplot(x)
}
