% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-helpers.R
\name{inline_hist}
\alias{inline_hist}
\title{Inline histogram}
\usage{
inline_hist(x, n_bins = 5)
}
\arguments{
\item{x}{Numeric vector.}

\item{n_bins}{Number of bins in the histogram.}
}
\value{
A compact character string with one bar per bin.
}
\description{
Inline histogram
}
\examples{
inline_hist(c(1, 2, 3, 4, 5))
x <- c(41, 53, 49, 52, NA, 47)
inline_hist(x)
}
