% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-helpers.R
\name{inline_scatter}
\alias{inline_scatter}
\title{Inline scatter plot}
\usage{
inline_scatter(x, y, width = 30, height = 8)
}
\arguments{
\item{x}{Numeric vector for the horizontal axis.}

\item{y}{Numeric vector for the vertical axis.}

\item{width}{Width of the plot in characters.}

\item{height}{Height of the plot in characters.}
}
\value{
A \code{birtir_inline_scatter} object.
}
\description{
Inline scatter plot
}
\examples{
inline_scatter(rnorm(50), rnorm(50))
}
