% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teikna_mynd.R
\name{lysa_stodu}
\alias{lysa_stodu}
\title{Deprecated legacy plot for student scale position}
\usage{
lysa_stodu(data, kvardi, fag)
}
\arguments{
\item{data}{einkunnir nemanda}

\item{kvardi}{tegund kvarda sem er notadur}

\item{fag}{"les" eda "stf" - raedur fjolda bakgrunsslita}
}
\value{
skilar mynd af stadsettningu nemanda
}
\description{
Deprecated legacy plot for student scale position
}
\details{
This function belongs to the legacy plotting workflow kept for reference
during the transition to \code{render_analysis_md()}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
lysa_stodu(fa_heildartolu(5), fa_kvarda(), fag = "les")
\dontshow{\}) # examplesIf}
}
