% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teikna_mynd_bekkjar.R
\name{lysa_stodu_bekkjar_legend}
\alias{lysa_stodu_bekkjar_legend}
\title{Mynd af legend fyrir lysa_stodu}
\usage{
lysa_stodu_bekkjar_legend()
}
\value{
skilar mynd af legend fyrir lysa_stodu
}
\description{
Mynd af legend fyrir lysa_stodu
}
\examples{
lysa_stodu_bekkjar_legend()

}
