% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjalpar_foll.R
\name{make_jitter}
\alias{make_jitter}
\title{Deprecated legacy helper for item-plot jitter}
\usage{
make_jitter(n)
}
\arguments{
\item{n}{Fjoldi gagnapunkta i hverri linu}
}
\value{
Laretta upprodun a gagnapunktum i atridagreining_einstaklinga()
}
\description{
Deprecated legacy helper for item-plot jitter
}
\details{
This function belongs to the legacy plotting workflow kept for reference
during the transition to \code{render_analysis_md()}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
make_jitter(unique(n))[point_index]
\dontshow{\}) # examplesIf}
}
