% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sanitize_name}
\alias{sanitize_name}
\title{Sanitize text for safe file/folder names}
\usage{
sanitize_name(x)
}
\arguments{
\item{x}{Character string.}
}
\value{
A cleaned string.
}
\description{
Sanitize text for safe file/folder names
}
\keyword{internal}
