% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-helpers.R
\name{as_report_text}
\alias{as_report_text}
\title{Convert a birtir description to report text}
\usage{
as_report_text(x, ...)
}
\arguments{
\item{x}{A \code{birtir_description} object.}

\item{...}{Reserved for future options.}
}
\value{
A single character string.
}
\description{
Convert a birtir description to report text
}
