% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teikna_mynd.R
\name{atridagreining_einstaklinga}
\alias{atridagreining_einstaklinga}
\title{Uppfærð mynd sem sýnir rétt/röng svör nemanda eftir atriðalýsingu og þyngd atriðis}
\usage{
atridagreining_einstaklinga(df)
}
\arguments{
\item{df}{einkunnir nemanda á atriðalevel}
}
\value{
skilar mynd af svörum nemenda eftir atriði og þyngd atriða
}
\description{
Uppfærð mynd sem sýnir rétt/röng svör nemanda eftir atriðalýsingu og þyngd atriðis
}
\examples{
atridagreining_einstaklinga(df)

}
