% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birtir-package.R
\docType{package}
\name{birtir-package}
\alias{birtir}
\alias{birtir-package}
\title{birtir: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/auv2/birtir}
  \item Report bugs at \url{https://github.com/auv2/birtir/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
