% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teikna_mynd.R
\name{einstaklingar_atridi}
\alias{einstaklingar_atridi}
\title{Mynd af atriðum nemanda - rétt/röng svör per atriði}
\usage{
einstaklingar_atridi(df_items)
}
\arguments{
\item{df_items}{einkunnir nemanda á atriðalevel}
}
\value{
skilar mynd af svörum nemenda eftir atriði og þyngd atriða
}
\description{
Mynd af atriðum nemanda - rétt/röng svör per atriði
}
\examples{
einstaklingar_atridi(df_items)
Áður en kallað er á fallið þarf að raða gögnunum með því að keyra þennan kóða þar sem df_atrdiði

# Raða x-ásnum og y-ásnum þannig að punktarnir flokkist rétt saman
df_student <- df_atridi \%>\%
  dplyr::mutate(
    difficulty = factor(difficulty, levels = c("Mjög létt", "Létt", "Þungt", "Mjög þungt"))
  ) \%>\%
  dplyr::arrange(difficulty, item_id) \%>\%  # Order by difficulty, then item_id
  mutate(student_response = factor(student_response, levels = c(TRUE, FALSE)))


df_student <- subset(df_student, nafn == "María Birna Karlsdóttir")

}
