% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjalpar_foll.R
\name{fa_heildartolu}
\alias{fa_heildartolu}
\title{litid hjalparfall sem byr til gogn}
\usage{
fa_heildartolu(maelitala)
}
\arguments{
\item{maelitala}{nemandans}
}
\value{
prufugogn
}
\description{
litid hjalparfall sem byr til gogn
}
\examples{
fa_heildartolu(10)
}
