% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teikna_mynd.R
\name{kortleggja}
\alias{kortleggja}
\title{Mynd af profil nemandans}
\usage{
kortleggja(data, kvardi)
}
\arguments{
\item{data}{einkunnir nemanda}

\item{kvardi}{tegund kvarda sem er notadur}
}
\value{
skilar mynd af stadsettningu nemanda
}
\description{
Mynd af profil nemandans
}
\examples{
kortleggja(fa_heildartolu(5), fa_kvarda())

}
