% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjalpar_foll.R
\name{litud_maelistika}
\alias{litud_maelistika}
\title{Byr til litada maelistiku}
\usage{
litud_maelistika(y_range, cutoffs, litur = "#D8C1FF", alpha = 1, fag = "les")
}
\arguments{
\item{y_range}{hæsta og lægsta gildi á kvarðanum}

\item{cutoffs}{thar sem skilin a milli flokka er}
\item{cutoffs}{þar sem skilin á milli flokka eru}

\item{litur}{litur kvarðans}

\item{alpha}{gegnsæi lita}

\item{fag}{fagið, "les" eða "stf" — ræður fjölda lita}
}
\value{
litadri mælistiku fyrir ggplot
}
\description{
Byr til litada maelistiku
}
