% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjalpar_foll.R
\name{litud_maelistika}
\alias{litud_maelistika}
\title{Byr til litada maelistiku}
\usage{
litud_maelistika(y_range, cutoffs, litur = "#C6D8CD", alpha = 0.8)
}
\arguments{
\item{y_range}{haesta og laegsta gildi a kvardanum}

\item{cutoffs}{thar sem skilin a milli flokka er}

\item{litur}{a kvardanum}

\item{alpha}{gildi sem lysir gegnsaei litarins}
}
\value{
litadri maelistiku fyrir ggplot
}
\description{
Byr til litada maelistiku
}
