% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teikna_mynd_bekkjar.R
\name{lysa_stodu_bekkjar}
\alias{lysa_stodu_bekkjar}
\title{Mynd af stadsetningu nemanda a kvarda asamt lysingu}
\usage{
lysa_stodu_bekkjar(data, kvardi, fag = "les")
}
\arguments{
\item{data}{einkunnir nemanda}

\item{kvardi}{tegund kvarda sem er notadur}

\item{fag}{les eða stf, ákvarðar fjölda lita í litud_maelistika()}
}
\value{
skilar mynd af stadsetningu bekkjar midad vid landsmedaltal a ollu profinu og undirthattum
}
\description{
Mynd af stadsetningu nemanda a kvarda asamt lysingu
}
\examples{
lysa_stodu_bekkjur(df, kvardi)

}
