% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjalpar_foll.R
\name{make_jitter}
\alias{make_jitter}
\title{Fall sem hjálpar að raða atriðunum í atridagreining_einstaklinga() fallinu}
\usage{
make_jitter(n)
}
\arguments{
\item{n}{Fjöldi gagnapunkta í hverri línu}
}
\value{
Lárétta uppröðun á gagnapunktum í atridagreining_einstaklinga()
}
\description{
Fall sem hjálpar að raða atriðunum í atridagreining_einstaklinga() fallinu
}
\examples{
make_jitter(unique(n))[point_index]
}
