% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teikna_mynd.R
\name{skifurit}
\alias{skifurit}
\title{Mynd sem sýnir hlutfall nemenda sem er undir ákveðnu viðmiði}
\usage{
skifurit(df, leidbeiningar, filter_expr)
}
\arguments{
\item{df}{einkunnir nemanda á atriðalevel með þyngd atriðanna flokkað í mjög létt, létt, þungt og mjög þungt}

\item{leidbeiningar}{umsögn um stöðu nemenda á viðmiðinu}

\item{filter_expr}{mælitölubilið sem verið er að skoða}
}
\value{
skilar mynd af hlutfalli nemenda á ákveðnu mælitölubili
}
\description{
Mynd sem sýnir hlutfall nemenda sem er undir ákveðnu viðmiði
}
\examples{
skifurit(df_bekkur, "þurfa mikinn stuðning", filter_expr < 3)

}
