% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teikna_mynd.R
\name{teikna_mynd}
\alias{teikna_mynd}
\title{Prenta ut myndir}
\usage{
teikna_mynd(data, kvardi)
}
\arguments{
\item{data}{einkunnir nemanda}

\item{kvardi}{tegund kvarda sem er notadur}
}
\value{
myndir fyrir nidurstodur
}
\description{
Prenta ut myndir
}
\examples{
data <- fa_heildartolu()
kvardi <- fa_kvarda()
teikna_mynd(data,kvardi)
}
