% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hjalpar_foll.R
\name{utbua_litapalletu}
\alias{utbua_litapalletu}
\title{Litapalleta}
\usage{
utbua_litapalletu(litur, fj_punkta)
}
\arguments{
\item{litur}{grunnlitur sem er sidan lystur upp med fj_punkta}

\item{fj_punkta}{fjoldi throskuldsgilda sem skipta kvardanum}
}
\value{
skilar kvarda med N litum
}
\description{
Litapalleta
}
\examples{
utbua_litapalletu("#C6D8CD", 4)
}
